examples/socket_python/requirements.txt
python-dotenv==1.0.1
supabase==2.11.0
